+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.11+2 (Backport)
+
+* Fix `spawnHybridUri` to respect language versioning of the spawned uri.
+
 ## 0.3.11+1
 
 * Allow analyzer 0.40.x.